BeanMachine 1.0
Added a camera for each scene that can be moved around as needed just by changing its position. The camera is also now what draws the scene.

Fixed some bugs like sprites colliding with ray casts and made massive changes to the sound manager as it was more broken than I thought.

Anyway the engine will now be used to hopefully make a simple game as I feel it is at a stage where that is possible.

Here's to BeanMachine 1.0!