Boiler Plate Mostly Done
Turns out the engine was far from ready to develop a game but hey its way better than it was before I started this game.

The camera has had a massive overhaul and now instead of scaling based on screen size its zooms in and out to try fit a set number of pixels in.

The engine now has integration with the TILED map editor so if you have a tiled json file and the tile sheet texture the engine will generate a map. it also supports animated tiles if you have multiple tile sheets with incrementing numbers.

And the last major thing is the UI module has finally been stared. However it wont be getting much more development as the Risdon and Risdon project will be entirely UI so that will probably do most of the work. The Ui in this game from now will probably be hacked together until the Risdon project has most of its Ui elements sorted.