Lighting Update
The Final Update before the massive engine revamp.

A new custom shape class has been created. When you pass in a list of vector2s it will create a shape with a texture you can use anywhere.

New addon called light. once added to  a sprite if lighting is enabled it will send its data to the lighting manager where it will be drawn over a layer of darkness.

Shadows are also supported but are very much experimental as of current. they use custom shapes and work well if only one light is casting shadows. however if 2 lights are casting it can cause issues.